#228 Refactored the InventorySales extension
