+++ app/code/Magento/InventoryCatalog/etc/di.xml
@@ -10,11 +10,11 @@
     <preference for="Magento\InventoryCatalog\Api\DefaultSourceProviderInterface" type="Magento\InventoryCatalog\Model\DefaultSourceProvider"/>
     <type name="Magento\InventoryApi\Api\StockRepositoryInterface">
         <plugin name="prevent_deleting_default_stock"
-                type="Magento\InventoryCatalog\Plugin\Inventory\Stock\PreventDeletingDefaultStockPlugin"/>
+                type="Magento\InventoryCatalog\Plugins\Inventory\Stock\PreventDeletingDefaultStockPlugin"/>
         <plugin name="prevent_deleting_stock_assigned_to_at_least_one_sales_channel"
-                type="Magento\InventoryCatalog\Plugin\Inventory\Stock\PreventDeletingStockWithSaleChannelAssignedPlugin"/>
+                type="Magento\InventoryCatalog\Plugins\Inventory\Stock\PreventDeletingStockWithSaleChannelAssignedPlugin"/>
     </type>
     <type name="Magento\InventoryApi\Api\ReservationsAppendInterface">
-        <plugin name="update_legacy_catalog_inventory_plugin" type="Magento\InventoryCatalog\Plugin\Inventory\Reservation\UpdateLegacyCatalogInventoryPlugin"/>
+        <plugin name="update_legacy_catalog_inventory_plugin" type="Magento\InventoryCatalog\Plugins\Inventory\Reservation\UpdateLegacyCatalogInventoryPlugin"/>
     </type>
 </config>
