+++ app/code/Magento/InventorySales/etc/di.xml
@@ -14,14 +14,14 @@
     <preference for="Magento\InventorySales\Model\GetAssignedStockIdForWebsiteInterface" type="Magento\InventorySales\Model\ResourceModel\GetAssignedStockIdForWebsite"/>
     <preference for="Magento\InventorySales\Model\DeleteSalesChannelToStockLinkInterface" type="Magento\InventorySales\Model\ResourceModel\DeleteSalesChannelToStockLink"/>
     <type name="Magento\InventoryApi\Api\StockRepositoryInterface">
-        <plugin name="load_sales_channels_on_get_list" type="Magento\InventorySales\Plugin\Inventory\StockRepository\LoadSalesChannelsOnGetListPlugin"/>
-        <plugin name="load_sales_channels_on_get" type="Magento\InventorySales\Plugin\Inventory\StockRepository\LoadSalesChannelsOnGetPlugin"/>
-        <plugin name="save_sales_channels_links" type="Magento\InventorySales\Plugin\Inventory\StockRepository\SaveSalesChannelsLinksPlugin"/>
+        <plugin name="enrich_stock_item_list_data_by_sale_channels_data" type="Magento\InventorySales\Plugins\Inventory\Stock\EnrichStockItemListDataBySalesChannelsDataPlugin"/>
+        <plugin name="enrich_stock_item_data_by_sale_channels_data" type="Magento\InventorySales\Plugins\Inventory\Stock\EnrichStockItemDataBySalesChannelsDataPlugin"/>
+        <plugin name="save_sales_channels_related_to_stock" type="Magento\InventorySales\Plugins\Inventory\Stock\SaveSalesChannelListRelatedToStockPlugin"/>
     </type>
     <type name="Magento\Inventory\Ui\DataProvider\StockDataProvider">
-        <plugin name="sales_channel_data" type="Magento\InventorySales\Plugin\Inventory\Ui\StockDataProvider\SalesChannels" />
+        <plugin name="enrich_stock_data_provider_by_sale_channel_data" type="Magento\InventorySales\Plugins\Inventory\Stock\EnrichStockDataProviderBySalesChannelDataPlugin" />
     </type>
     <type name="Magento\InventoryApi\Api\IsProductInStockInterface">
-        <plugin name="backorder_inventory_stock_status" type="Magento\InventorySales\Plugin\InventoryApi\BackorderStockStatusPlugin"/>
+        <plugin name="check_ability_to_backorder_stock_items" type="Magento\InventorySales\Plugins\Inventory\Stock\CheckAbilityToBackorderStockItemPlugin"/>
     </type>
 </config>
